Loaded all textures in onInit.
