+++ gvrf_accessibility_sample/src/com/samsung/accessibility/shortcut/ShortcutMenuItem.java
@@ -1,9 +1,4 @@
-
 package com.samsung.accessibility.shortcut;
-
-import com.samsung.accessibility.R;
-import com.samsung.accessibility.focus.FocusableSceneObject;
-import com.samsung.accessibility.focus.OnFocusListener;
 
 import org.gearvrf.GVRAndroidResource;
 import org.gearvrf.GVRContext;
@@ -14,6 +9,11 @@ import org.gearvrf.GVRRenderData.GVRRenderingOrder;
 import org.gearvrf.GVRSceneObject;
 import org.gearvrf.GVRTexture;
 
+import com.samsung.accessibility.R;
+import com.samsung.accessibility.focus.FocusableSceneObject;
+import com.samsung.accessibility.focus.OnFocusListener;
+import com.samsung.accessibility.util.AccessibilityTexture;
+
 public class ShortcutMenuItem extends FocusableSceneObject {
 
     private GVRContext gvrContext;
@@ -21,16 +21,6 @@ public class ShortcutMenuItem extends FocusableSceneObject {
     private static final int LOST_FOCUS_COLOR = 6186095;
     private static final int CLICKED_COLOR = 12631476;
     private boolean clicked;
-    private GVRTexture accessibilityIcon;
-    private GVRTexture backIcon;
-    private GVRTexture spaceTexture;
-    private GVRTexture zoomOut;
-    private GVRTexture zoomIn;
-    private GVRTexture talkBackLess;
-    private GVRTexture talkBackMore;
-    private GVRTexture invertedColorsIcon;
-    private GVRTexture emptyIcon;
-    private GVRTexture speechIcon;
     private GVRSceneObject icon;
     private TypeItem typeItem;
 
@@ -40,7 +30,6 @@ public class ShortcutMenuItem extends FocusableSceneObject {
         createRenderData();
         attachEyePointeeHolder();
         getRenderData().getMaterial().setColor(LOST_FOCUS_COLOR);
-        loadFiles();
     }
 
     private void createRenderData() {
@@ -52,19 +41,6 @@ public class ShortcutMenuItem extends FocusableSceneObject {
         renderData.setMesh(mesh);
         attachRenderData(renderData);
         getRenderData().getMaterial().setMainTexture(texture);
-    }
-
-    private void loadFiles() {
-        accessibilityIcon = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_accessibility));
-        backIcon = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_back));
-        spaceTexture = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.circle_normal_alpha));
-        talkBackMore = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_talkback_mais));
-        talkBackLess = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_talkback_menos));
-        zoomIn = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_zoom_mais));
-        invertedColorsIcon = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_inverted));
-        zoomOut = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_zoom_menos));
-        speechIcon = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.ico_speech));
-        emptyIcon = gvrContext.loadTexture(new GVRAndroidResource(gvrContext, R.drawable.empty));
     }
 
     public void focusAndUnFocus() {
@@ -102,14 +78,18 @@ public class ShortcutMenuItem extends FocusableSceneObject {
         icon.getTransform().rotateByAxis(-90, 1, 0, 0);
         icon.getTransform().rotateByAxisWithPivot(245, 0, 1, 0, 0, 0, 0);
         icon.getRenderData().setRenderingOrder(GVRRenderingOrder.OVERLAY);
-        getRenderData().getMaterial().setMainTexture(spaceTexture);
+        getRenderData().getMaterial().setMainTexture(AccessibilityTexture.getInstance(gvrContext).getSpaceTexture());
         this.typeItem = typeItem;
         addChildObject(icon);
     }
 
     public void removeIcon() {
         typeItem = TypeItem.EMPTY;
-        icon.getRenderData().getMaterial().setMainTexture(emptyIcon);
+        icon.getRenderData().getMaterial().setMainTexture(AccessibilityTexture.getInstance(gvrContext).getEmptyIcon());
+    }
+
+    public GVRSceneObject getIcon() {
+        return icon;
     }
 
     public void setClicked(boolean clicked) {
@@ -126,46 +106,6 @@ public class ShortcutMenuItem extends FocusableSceneObject {
 
     public void setTypeItem(TypeItem typeItem) {
         this.typeItem = typeItem;
-    }
-
-    public GVRTexture getAccessibilityIcon() {
-        return accessibilityIcon;
-    }
-
-    public GVRTexture getBackIcon() {
-        return backIcon;
-    }
-
-    public GVRTexture getZoomOut() {
-        return zoomOut;
-    }
-
-    public GVRTexture getZoomIn() {
-        return zoomIn;
-    }
-
-    public GVRTexture getTalkBackLess() {
-        return talkBackLess;
-    }
-
-    public GVRTexture getTalkBackMore() {
-        return talkBackMore;
-    }
-
-    public GVRTexture getInvertedColorsIcon() {
-        return invertedColorsIcon;
-    }
-
-    public GVRTexture getSpeechIcon() {
-        return speechIcon;
-    }
-
-    public GVRSceneObject getIcon() {
-        return icon;
-    }
-
-    public GVRTexture getEmptyIcon() {
-        return emptyIcon;
     }
 
     public enum TypeItem {
